feat: refactoring of code
